Refactor : Habit 전체 검색 API 리팩토링
