+++ app/src/repository/HabitRepository.ts
@@ -1,4 +1,5 @@
-import { HabitCreateInput, PrismaClient } from '@prisma/client';
+import { HabitCreateInput, PrismaClient, User } from '@prisma/client';
+import moment from 'moment';
 import { BaseRepository } from './BaseRepository';
 
 export class HabitRepository extends BaseRepository {
@@ -12,6 +13,27 @@ export class HabitRepository extends BaseRepository {
   public async createHabitJoinUser(payload: HabitCreateInput) {
     return this.prisma.habit.create({
       data: payload,
+    });
+  }
+
+  public async findAllHabitWithinAWeekByUserId(user: User) {
+    return this.prisma.habit.findMany({
+      where: { userId: user.userId },
+      select: {
+        habitId: true,
+        title: true,
+        description: true,
+        dayOfWeek: true,
+        commitHistory: {
+          where: {
+            createdAt: {
+              gte: moment().startOf('weeks').toDate(),
+              lte: moment().endOf('weeks').toDate(),
+            },
+          },
+          select: { createdAt: true },
+        },
+      },
     });
   }
 }
